# swagger.yaml
swagger: "2.0"
info:
  description: "Zum Zählen von Punkten beim Jassen"
  version: "1.0.0"
  title: "Schieber Tafel"
  termsOfService: ""
  contact:
    email: "jass@scheissegal.io"
  license:
    name: "Apache 2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
host: "jass.scheissegal.io"
basePath: "/v1"
...
schemes:
  - "https"
  - "http"
paths:
  /team:
    get:
      tags:
        - "team"
      summary: "get all teams"
      description: ""
      operationId: "getTeam"
      produces:
        - "application/json"
      responses:
        200:
          description: "All teams"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Team"
    post:
      tags:
        - "team"
      summary: "create new team"
      description: ""
      operationId: "createTeam"
      consumes:
        - "application/json"
      produces:
        - "application/json"
      parameters:
        - in: "body"
          name: "body"
          description: "Team to create"
          required: true
          schema:
            $ref: "#/definitions/Team"
      responses:
        201:
          description: "Team created"
          schema:
            type: "object"
            properties:
              id:
                type: "integer"
    put:
      tags:
        - "team"
      summary: "update team"
      description: ""
      operationId: "updateTeam"
      consumes:
        - "application/json"
      produces:
        - "application/json"
      parameters:
        - in: "body"
          name: "body"
          description: "Team to update"
          required: true
          schema:
            $ref: "#/definitions/Team"
      responses:
        200:
          description: "Team updated"
...
definitions:
  Team:
    type: "object"
    properties:
      id:
        type: "integer"
        format: "int64"
      name:
        type: "string"
...